Actions: make milestone detection available for all tasks
